Tutor/backend/pyproject.toml
[project]
name = "tutorgpt-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiosqlite>=0.21.0",
    "chromadb>=1.3.4",
    "fastapi>=0.121.1",
    "google-genai>=1.49.0",
    "openai-agents>=0.5.0",
    "pydantic>=2.12.4",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "pytest-cov>=7.0.0",
    "python-dotenv>=1.2.1",
    "python-frontmatter>=1.0.0",
    "tqdm>=4.66.0",
    "uvicorn[standard]>=0.38.0",
]
